postgres

+++ project/__init__.py
@@ -2,7 +2,7 @@ import logging
 from logging.handlers import RotatingFileHandler
 import os
 
-# import sqlalchemy as sa
+import sqlalchemy as sa
 from flask import Flask
 from flask.logging import default_handler
 from flask_sqlalchemy import SQLAlchemy
@@ -21,17 +21,17 @@ def create_app():
     register_blueprints(app)
     configure_logging(app)
 
-    # engine = sa.create_engine(app.config['SQLALCHEMY_DATABASE_URI'])
-    # inspector = sa.inspect(engine)
-    # if not inspector.has_table("todo"):
-    with app.app_context():
+    engine = sa.create_engine(app.config['SQLALCHEMY_DATABASE_URI'])
+    inspector = sa.inspect(engine)
+    if not inspector.has_table("todo"):
+        with app.app_context():
+            db.init_app(app)
+            db.drop_all()
+            db.create_all()
+            app.logger.info('Initialized the database!')
+    else:
         db.init_app(app)
-        db.drop_all()
-        db.create_all()
-        app.logger.info('Initialized the database!')
-    # else:
-    #    db.init_app(app)
-    #    app.logger.info('Database already contains the users table.')
+        app.logger.info('Database already contains the users table.')
 
     return app
